server/requirements.txt
boto3==1.35.63
botocore==1.35.63
build==1.1.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
docopt==0.6.2
greenlet==2.0.2
idna==3.10
importlib-metadata==6.7.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.2
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pip-tools==6.14.0
pipreqs==0.4.13
playwright==1.35.0
pyee==9.0.4
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.5
requests==2.31.0
s3transfer==0.10.3
six==1.17.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.20
yarg==0.1.10
zipp==3.15.0
flask==3.0.1
flask-cors==5.0.1
